--- src/sources/kr/Agitoon.ts
@@ -1,32 +1,0 @@
-import { fetchHtml } from '@utils/fetch/fetch';
-import * as cheerio from 'cheerio';
-import { SourceChapter, SourceNovel, SourceNovelItem } from '../types';
-
-const sourceId = 145;
-const sourceName = 'Agitoon';
-const baseUrl = 'https://agit501.xyz/';
-
-const popularNovels = () => {
-  return null;
-};
-
-const parseNovelAndChapters = () => {
-  return null;
-};
-
-const parseChapter = () => {
-  return null;
-};
-
-const searchNovels = () => {
-  return null;
-};
-
-const AgitoonScraper = {
-  popularNovels,
-  parseNovelAndChapters,
-  parseChapter,
-  searchNovels,
-};
-
-export default AgitoonScraper;
